move subcase functions to xpath_functions package
