+++ corehq/apps/case_search/xpath_functions/__init__.py
@@ -1,6 +1,12 @@
+from .subcase_functions import subcase
 from .value_functions import date, XPathFunctionException  # noqa: F401
 
 # functions that transform a value
 XPATH_VALUE_FUNCTIONS = {
     'date': date,
+}
+
+
+XPATH_QUERY_FUNCTIONS = {
+    'subcase-exists': subcase,
 }
